// packages/nodejs/codepot/src/contract/schema/request-bodies/request-body-component.types.ts
import type { ComponentRef, ModelRef, RequestBodyRef } from '../../refs/ref.types';
import type { EngineRef } from '../../refs/ref.types';
import type { RefUsage } from '../../refs/ref-usage.types';
import { CodegenMetadata } from '@/pipeline/targets/codegen/codegen-extension.types';
import { ComponentFieldMap } from '@/pipeline/targets/openapi/components/component.types';

export type RequestBodySchemaRef = ComponentRef | ModelRef | ComponentFieldMap | RefUsage<EngineRef>;

export interface RequestBodyComponentDefinition {
  readonly name: string;
  readonly schema: RequestBodySchemaRef;
  readonly required?: boolean;
  readonly description?: string;
  readonly meta?: CodegenMetadata;
}

export interface RequestBodyComponentRegistry<TRefs extends Record<string, RequestBodyRef> = Record<string, RequestBodyRef>> {
  readonly name: string;
  readonly definitions: RequestBodyComponentDefinition[];
  readonly ref: TRefs;
}